refactor: COCO moved to the legacy directory
