+++ convert/src/badgerdoc_format/badgerdoc_format.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import Optional
 
-from ...config import (
+from ..config import (
     DEFAULT_PAGE_BORDER_OFFSET,
     DEFAULT_PDF_FONT_HEIGHT,
     DEFAULT_PDF_FONT_WIDTH,
@@ -10,7 +10,7 @@ from ...config import (
 )
 from .bd_annotation_model_practic import BadgerdocAnnotation
 from .bd_tokens_model import Page
-from src.labelstudio_to_badgerdoc.labelstudio_format.labelstudio_models import LabelStudioModel
+from src.labelstudio_format.labelstudio_models import LabelStudioModel
 from .annotation_converter import AnnotationConverter
 from .pdf_renderer import PDFRenderer
 from .plain_text_converter import TextToBadgerdocTokensConverter
