+++ convert/src/legacy/coco_import/convert.py
@@ -7,9 +7,9 @@ from typing import Any, Dict, Set
 
 from src.config import get_request_session, settings
 from src.logger import get_logger
-from src.models.coco import DataS3
-from src.utils.json_utils import import_save_to_json, load_from_json
-from src.utils.s3_utils import S3Manager, s3_download_files
+from ..models.coco import DataS3
+from ..utils.json_utils import import_save_to_json, load_from_json
+from ..utils.s3_utils import S3Manager, s3_download_files
 
 LOGGER = get_logger(__file__)
 SESSION = get_request_session()
