+++ convert/src/legacy/exceptions.py
