+++ convert/src/legacy/utils/render_pdf_page.py
@@ -6,7 +6,7 @@ import pdfplumber
 
 from src.config import settings
 from src.logger import get_logger
-from src.utils.common_utils import add_to_zip_and_local_remove
+from .common_utils import add_to_zip_and_local_remove
 
 LOGGER = get_logger(__file__)
