+++ convert/src/legacy/utils/s3_utils.py
@@ -6,10 +6,10 @@ import urllib3
 from fastapi import HTTPException, status
 
 from src.config import settings
-from src.exceptions import BucketError, FileKeyError, UploadLimitExceedError
+from ..exceptions import BucketError, FileKeyError, UploadLimitExceedError
 from src.logger import get_logger
-from src.models import coco
-from src.utils.common_utils import check_uploading_limit
+from src.legacy.models import coco
+from src.legacy.utils.common_utils import check_uploading_limit
 
 logger = get_logger(__name__)
