+++ convert/src/text_to_badgerdoc_converter.py
@@ -1,7 +1,7 @@
 import tempfile
 from pathlib import Path
 
-from ..config import DEFAULT_PAGE_BORDER_OFFSET
+from .config import DEFAULT_PAGE_BORDER_OFFSET
 from .badgerdoc_format.badgerdoc_format import BadgerdocFormat
 from .badgerdoc_format.pdf_renderer import PDFRenderer
 from .badgerdoc_format.plain_text_converter import (
